requirements.txt
python-dotenv
langchain
langchain-anthropic
langchain-community
langchain-core
langchain-openai
langchain-chroma
langchain-huggingface
streamlit
pdfminer.six
unstructured[all-docs]